@@ -28,16 +28,16 @@
     "@storybook/manager-webpack4": "^6.5.6",
     "@storybook/react": "^6.5.6",
     "@storybook/testing-library": "^0.0.11",
-    "@types/react": "*",
-    "@types/react-dom": "^18.0.5",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "@types/react-modal": "^3.13.1",
     "babel-loader": "^8.2.5",
     "classnames": "^2.3.1",
     "node-typescript": "^0.1.3",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "require-from-string": "^2.0.2",
-    "type-route": "^0.7.1",
+    "type-route": "^1.0.0",
     "typescript": ">=4.3.0",
     "webpack": ">=2.0.0"
   },
