@@ -1,20 +1,20 @@
 import "src/assets/css/index.css";
 import "src/assets/dsfr/dsfr.min.css";
 import "src/assets/dsfr/utility/icons/icons.min.css";
-
 import React from "react";
-import ReactDOM from "react-dom";
+import { startReactDsfr } from "@codegouvfr/react-dsfr/spa";
+import { createRoot } from "react-dom/client";
 import { Provider } from "react-redux";
 import { App } from "src/app/App";
 import { store } from "src/config/dependencies";
 import { MetaContent } from "./components/layout/MetaContent";
-
-import { startReactDsfr } from "@codegouvfr/react-dsfr/spa";
-
-startReactDsfr({ defaultColorScheme: "light" });
 import { RouteProvider } from "./routes/routes";
 
-ReactDOM.render(
+startReactDsfr({ defaultColorScheme: "light" });
+
+const rootContainer = document.getElementById("root");
+if (!rootContainer) throw new Error("Html Element with Id 'root' is missing.");
+createRoot(rootContainer).render(
   <React.StrictMode>
     <Provider store={store}>
       <RouteProvider>
@@ -23,5 +23,4 @@ ReactDOM.render(
       </RouteProvider>
     </Provider>
   </React.StrictMode>,
-  document.getElementById("root"),
 );
