@@ -182,9 +182,10 @@ describe("PgConventionPoleEmploiAdvisorRepository", () => {
     it("should return undefined if no convention Advisor", async () => {
       const conventionAdvisor:
         | ConventionPoleEmploiUserAdvisorEntity
-        | undefined = await conventionPoleEmploiAdvisorRepository.getByConventionId(
-        conventionId,
-      );
+        | undefined =
+        await conventionPoleEmploiAdvisorRepository.getByConventionId(
+          conventionId,
+        );
 
       expect(conventionAdvisor).toBeUndefined();
     });
@@ -200,9 +201,10 @@ describe("PgConventionPoleEmploiAdvisorRepository", () => {
 
       const conventionAdvisor:
         | ConventionPoleEmploiUserAdvisorEntity
-        | undefined = await conventionPoleEmploiAdvisorRepository.getByConventionId(
-        conventionId,
-      );
+        | undefined =
+        await conventionPoleEmploiAdvisorRepository.getByConventionId(
+          conventionId,
+        );
 
       expectObjectsToMatch(conventionAdvisor!, {
         advisor: poleEmploiFirstUserAdvisor.advisor,
@@ -221,9 +223,10 @@ describe("PgConventionPoleEmploiAdvisorRepository", () => {
 
       const conventionAdvisor:
         | ConventionPoleEmploiUserAdvisorEntity
-        | undefined = await conventionPoleEmploiAdvisorRepository.getByConventionId(
-        conventionId,
-      );
+        | undefined =
+        await conventionPoleEmploiAdvisorRepository.getByConventionId(
+          conventionId,
+        );
 
       expectObjectsToMatch(conventionAdvisor!, {
         advisor: undefined,
